update function login, forgot and reset password

+++ ManagerTour/Controllers/TourController.cs
@@ -32,105 +32,142 @@ namespace ManagerTour.Controllers
         }
 
         // GET: Tour
-        public ActionResult Index(string keyword = null)
+        public ActionResult Index(string keyword = null, string filter_status = "0")
         {
-            string query = "SELECT t.id, t.user_id, t.vehicle_id, t.title, t.description, t.address_start, t.address_end, t.date_start, t.date_end, " +
-                            "t.price_tour, t.detail_price_tour, t.amount_customer_maximum, t.amount_customer_present, t.status, t.created_at, u.name as userName, v.name as vehicleName, t.created_at, t.updated_at" +
-                            " FROM tours as t join user_information as u on t.user_id = u.id join vehicles as v on t.vehicle_id = v.id";
-
-            //if keyword search exists perform search with field title or description of the tour
-            if(!string.IsNullOrEmpty(keyword) && !string.IsNullOrWhiteSpace(keyword))
+            if (Session["user"] != null)
             {
-                query += " WHERE title like '%" + keyword + "%' or description like '%" + keyword + "%'";
-            }
-
-
-            int totalRecords = (currentPage - 1) * pageSize;
-
-            query += " LIMIT " + pageSize + " OFFSET " + totalRecords;
-
-
-            ConnectionMySQL connect = new ConnectionMySQL();
-            DataTable dt = new DataTable();
-            dt = connect.SelectData(query).Tables[0];
-
-            if(dt.Rows.Count > 0)
-            {
-                foreach(DataRow row in dt.Rows)
+                try
                 {
-                    Tours tour = new Tours
-                    {
-                        Id = Int32.Parse(row["id"].ToString()),
-                        User_id = Int32.Parse(row["user_id"].ToString()),
-                        Vehicle_id = Int32.Parse(row["vehicle_id"].ToString()),
-                        Title = row["title"].ToString(),
-                        Description = row["description"].ToString(),
-                        Address_start = row["address_start"].ToString(),
-                        Address_end = row["address_end"].ToString(),
-                        Date_start = String.Format("{0:yyyy-MM-dd}", row["date_start"]),
-                        Date_end = String.Format("{0:yyyy-MM-dd}", row["date_end"]),
-                        Price_tour = double.Parse(row["price_tour"].ToString()),
-                        Detail_price_tour = row["detail_price_tour"].ToString(),
-                        Amount_customer_maximum = Int32.Parse(row["amount_customer_maximum"].ToString()),
-                        Amount_customer_present = Int32.Parse(row["amount_customer_present"].ToString()),
-                        Status = Int32.Parse(row["status"].ToString()),
-                        Created_at = String.Format("{0:dd-MM-yyyy}", row["created_at"].ToString()),
-                        Updated_at = String.Format("{0:dd-MM-yyyy}", row["updated_at"].ToString()),
-                        User = new User_information { Name = row["userName"].ToString() },
-                        Vehicles = new Vehicles { Name = row["vehicleName"].ToString() },
-                        TotalPage = totalPage,
-                        CurrentPage = currentPage,
-                    };
 
-                    ListTour.Add(tour);
+                    string query = "SELECT t.id, t.user_id, t.vehicle_id, t.title, t.description, t.address_start, t.address_end, t.date_start, t.date_end, " +
+                                "t.price_tour, t.detail_price_tour, t.amount_customer_maximum, t.amount_customer_present, t.status, t.created_at, u.name as userName, v.name as vehicleName, t.created_at, t.updated_at" +
+                                " FROM tours as t join user_information as u on t.user_id = u.id join vehicles as v on t.vehicle_id = v.id";
+
+                    if (!string.IsNullOrEmpty(filter_status) && !string.IsNullOrWhiteSpace(filter_status))
+                    {
+                        switch (filter_status)
+                        {
+                            case "1":
+                                query += " WHERE t.status = 0";
+                                break;
+                            case "2":
+                                query += " WHERE t.status = 1";
+                                break;
+                            case "3":
+                                query += " WHERE t.status = 2";
+                                break;
+                            default:
+                                query += " WHERE 1";
+                                break;
+                        }
+                    }
+
+                    //if keyword search exists perform search with field title or description of the tour
+                    if (!string.IsNullOrEmpty(keyword) && !string.IsNullOrWhiteSpace(keyword))
+                    {
+                        query += " and title like '%" + keyword + "%' or description like '%" + keyword + "%'";
+                    }
+
+                    int totalRecords = (currentPage - 1) * pageSize;
+
+                    query += " ORDER BY t.id LIMIT " + pageSize + " OFFSET " + totalRecords;
+
+
+                    ConnectionMySQL connect = new ConnectionMySQL();
+                    DataTable dt = new DataTable();
+                    dt = connect.SelectData(query).Tables[0];
+
+                    if (dt.Rows.Count > 0)
+                    {
+                        foreach (DataRow row in dt.Rows)
+                        {
+                            Tours tour = new Tours
+                            {
+                                Id = Int32.Parse(row["id"].ToString()),
+                                User_id = Int32.Parse(row["user_id"].ToString()),
+                                Vehicle_id = Int32.Parse(row["vehicle_id"].ToString()),
+                                Title = row["title"].ToString(),
+                                Description = row["description"].ToString(),
+                                Address_start = row["address_start"].ToString(),
+                                Address_end = row["address_end"].ToString(),
+                                Date_start = String.Format("{0:yyyy-MM-dd}", row["date_start"]),
+                                Date_end = String.Format("{0:yyyy-MM-dd}", row["date_end"]),
+                                Price_tour = double.Parse(row["price_tour"].ToString()),
+                                Detail_price_tour = row["detail_price_tour"].ToString(),
+                                Amount_customer_maximum = Int32.Parse(row["amount_customer_maximum"].ToString()),
+                                Amount_customer_present = Int32.Parse(row["amount_customer_present"].ToString()),
+                                Status = Int32.Parse(row["status"].ToString()),
+                                Created_at = String.Format("{0:dd-MM-yyyy}", row["created_at"].ToString()),
+                                Updated_at = String.Format("{0:dd-MM-yyyy}", row["updated_at"].ToString()),
+                                User = new User_information { Name = row["userName"].ToString() },
+                                Vehicles = new Vehicles { Name = row["vehicleName"].ToString() },
+                                TotalPage = totalPage,
+                                CurrentPage = currentPage,
+                            };
+
+                            ListTour.Add(tour);
+                        }
+                    }
+
+                    ViewBag.keyword = keyword;
+                    ViewBag.filter_status = filter_status;
+
+                    return View(ListTour);
+                }
+                catch
+                {
+
                 }
             }
 
-            ViewBag.keyword = keyword;
-
-            return View(ListTour);
+            return RedirectToAction("Login", "Auth");
         }
 
         //get view detail of the tour
         public ActionResult Detail(int id)
         {
-            string query = "SELECT * FROM tours WHERE id = " + id;
-            ConnectionMySQL connect = new ConnectionMySQL();
-            DataTable dt = new DataTable();
-            dt = connect.SelectData(query).Tables[0];
-
-            if (dt.Rows.Count > 0)
+            if (Session["user"] != null)
             {
-                foreach (DataRow row in dt.Rows)
+                string query = "SELECT * FROM tours WHERE id = " + id;
+                ConnectionMySQL connect = new ConnectionMySQL();
+                DataTable dt = new DataTable();
+                dt = connect.SelectData(query).Tables[0];
+
+                if (dt.Rows.Count > 0)
                 {
-                    Tours tour = new Tours
+                    foreach (DataRow row in dt.Rows)
                     {
-                        Id = Int32.Parse(row["id"].ToString()),
-                        User_id = Int32.Parse(row["user_id"].ToString()),
-                        Vehicle_id = Int32.Parse(row["vehicle_id"].ToString()),
-                        Title = row["title"].ToString(),
-                        Description = row["description"].ToString(),
-                        Address_start = row["address_start"].ToString(),
-                        Address_end = row["address_end"].ToString(),
-                        Date_start = String.Format("{0:yyyy-MM-dd}", row["date_start"]),
-                        Date_end = String.Format("{0:yyyy-MM-dd}", row["date_end"]),
-                        Price_tour = double.Parse(row["price_tour"].ToString()),
-                        Detail_price_tour = row["detail_price_tour"].ToString(),
-                        Amount_customer_maximum = Int32.Parse(row["amount_customer_maximum"].ToString()),
-                        Amount_customer_present = Int32.Parse(row["amount_customer_present"].ToString()),
-                        Status = Int32.Parse(row["status"].ToString()),
-                        Created_at = String.Format("{0:dd-MM-yyyy}", row["created_at"].ToString()),
-                        Updated_at = String.Format("{0:dd-MM-yyyy}", row["updated_at"].ToString()),
-                        ListVehicle = ListVehicle,
-                    };
+                        Tours tour = new Tours
+                        {
+                            Id = Int32.Parse(row["id"].ToString()),
+                            User_id = Int32.Parse(row["user_id"].ToString()),
+                            Vehicle_id = Int32.Parse(row["vehicle_id"].ToString()),
+                            Title = row["title"].ToString(),
+                            Description = row["description"].ToString(),
+                            Address_start = row["address_start"].ToString(),
+                            Address_end = row["address_end"].ToString(),
+                            Date_start = String.Format("{0:yyyy-MM-dd}", row["date_start"]),
+                            Date_end = String.Format("{0:yyyy-MM-dd}", row["date_end"]),
+                            Price_tour = double.Parse(row["price_tour"].ToString()),
+                            Detail_price_tour = row["detail_price_tour"].ToString(),
+                            Amount_customer_maximum = Int32.Parse(row["amount_customer_maximum"].ToString()),
+                            Amount_customer_present = Int32.Parse(row["amount_customer_present"].ToString()),
+                            Status = Int32.Parse(row["status"].ToString()),
+                            Created_at = String.Format("{0:dd-MM-yyyy}", row["created_at"].ToString()),
+                            Updated_at = String.Format("{0:dd-MM-yyyy}", row["updated_at"].ToString()),
+                            ListVehicle = ListVehicle,
+                        };
 
 
-                    ListTour.Add(tour);
+                        ListTour.Add(tour);
+                    }
+
                 }
 
+                return View(ListTour);
             }
 
-            return View(ListTour);
+            return RedirectToAction("Login", "Auth");
         }
 
         //get list Address
@@ -181,47 +218,87 @@ namespace ManagerTour.Controllers
         //delete tour by ID
         public ActionResult Delete(int id)
         {
-            try
+            if (Session["user"] != null)
             {
-                string query = "DELETE FROM `tours` WHERE id = " + id;
-                ConnectionMySQL connect = new ConnectionMySQL();
-                connect.ExecuteNonQuery(query);
+                try
+                {
+                    string query = "DELETE FROM `tours` WHERE id = " + id;
+                    ConnectionMySQL connect = new ConnectionMySQL();
+                    connect.ExecuteNonQuery(query);
 
-                return RedirectToAction("Index");
+                    return RedirectToAction("Index");
+                }
+                catch (Exception e)
+                {
+                    return RedirectToAction("Index");
+                }
             }
-            catch (Exception e)
-            {
-                return RedirectToAction("Index");
-            }
+
+            return RedirectToAction("Login", "Auth");
         }
 
 
         //update information of the tour 
         public ActionResult Update(int id, int status_tour, List<Tours> tour)
         {
-            try
+            if (Session["user"] != null)
             {
-                if (ModelState.IsValid)
+                try
                 {
-                    string query = "UPDATE `tours` SET `title`='"+tour[0].Title+ "',`description`='" + tour[0].Description + "',`price_tour`='" + tour[0].Price_tour + "'," +
-                        "`detail_price_tour`='" + tour[0].Detail_price_tour + "',`amount_customer_maximum`='" + tour[0].Amount_customer_maximum + "', `amount_customer_present`='" + tour[0].Amount_customer_present + "'," +
-                        "`status`='" + status_tour + "',`updated_at`='"+DateTime.Now.ToString("yyyy-MM-dd H:m:s")+"'" +
-                        " WHERE id = " + id;
+                    if (ModelState.IsValid)
+                    {
+                        string query = "UPDATE `tours` SET `title`='" + tour[0].Title + "',`description`='" + tour[0].Description + "',`price_tour`='" + tour[0].Price_tour + "'," +
+                            "`detail_price_tour`='" + tour[0].Detail_price_tour + "',`amount_customer_maximum`='" + tour[0].Amount_customer_maximum + "', `amount_customer_present`='" + tour[0].Amount_customer_present + "'," +
+                            "`status`='" + status_tour + "',`updated_at`='" + DateTime.Now.ToString("yyyy-MM-dd H:m:s") + "'" +
+                            " WHERE id = " + id;
 
-                    ConnectionMySQL connect = new ConnectionMySQL();
-                    connect.ExecuteNonQuery(query);
+                        ConnectionMySQL connect = new ConnectionMySQL();
+                        connect.ExecuteNonQuery(query);
 
-                    return RedirectToAction("Index");
+                        return RedirectToAction("Index");
+                    }
+                    else
+                    {
+                        return RedirectToAction("Detail");
+                    }
                 }
-                else
+                catch (Exception e)
                 {
                     return RedirectToAction("Detail");
                 }
             }
-            catch (Exception e)
+
+            return RedirectToAction("Login", "Auth");
+        }
+
+        //agree tour
+        public ActionResult AgreeTour(int id)
+        {
+            if (Session["user"] != null)
             {
-                return RedirectToAction("Detail");
+                try
+                {
+                    if (ModelState.IsValid)
+                    {
+                        string query = "UPDATE `tours` SET `status`='1', `updated_at`='" + DateTime.Now.ToString("yyyy-MM-dd H:m:s") + "' WHERE id = " + id;
+
+                        ConnectionMySQL connect = new ConnectionMySQL();
+                        connect.ExecuteNonQuery(query);
+
+                        return RedirectToAction("Index");
+                    }
+                    else
+                    {
+                        return RedirectToAction("Detail");
+                    }
+                }
+                catch (Exception e)
+                {
+                    return RedirectToAction("Detail");
+                }
             }
+
+            return RedirectToAction("Login", "Auth");
         }
 
         //pagination next page
